[feat](linked_list): reverse_list
adding function that will reverse the list, it first init with 3 Node
pointers: previous, current and next.

previous and next init with NULL, at every iteration next holds the
value of the next pointer within current.
then update the value of this next pointer within current with the value
of the previous
then we update the value of the previous with the current value
and last the current value is updated to use the next element captured
before